End-to-end preview XML generated

+++ app/helpers/metadata_builder_helper.rb
@@ -30,15 +30,17 @@ module MetadataBuilderHelper
     @sample_xml_docs = ""
     @file_links = Array.new
     @object.metadata_builder.preserve.each do |file|
-      @file_links << link_to(prettify(file), "##{file}")
-      anchor_tag = content_tag(:a, "", :name=> file)
-      sample_xml_content = File.open(file, "r"){|io| io.read}
-      sample_xml_doc = REXML::Document.new sample_xml_content
-      sample_xml = ""
-      sample_xml_doc.write(sample_xml, 1)
-      header = content_tag(:h3, "XML Sample for #{prettify(file)}")
-      xml_code = content_tag(:pre, "#{sample_xml}")
-      @sample_xml_docs << content_tag(:div, anchor_tag << header << xml_code, :class => "doc")
+      if File.exist?(file)
+        @file_links << link_to(prettify(file), "##{file}")
+        anchor_tag = content_tag(:a, "", :name=> file)
+        sample_xml_content = File.open(file, "r"){|io| io.read}
+        sample_xml_doc = REXML::Document.new sample_xml_content
+        sample_xml = ""
+        sample_xml_doc.write(sample_xml, 1)
+        header = content_tag(:h3, "XML Sample for #{prettify(file)}")
+        xml_code = content_tag(:pre, "#{sample_xml}")
+        @sample_xml_docs << content_tag(:div, anchor_tag << header << xml_code, :class => "doc")
+      end
     end
     @object.version_control_agent.delete_clone
     @file_links_html = ""
